【step09】实现解析编译后的材质脚本

+++ source/step09/Core/Include/Kernel/T3DCommon.h
@@ -787,7 +787,7 @@ namespace Tiny3D
         char        magic[4];   /**< 文件魔数，用于标识文件 */
         uint32_t    subtype;    /**< The subtype */
         uint32_t    version;    /**< 文件版本号 */
-        size_t      filesize;   /**< 文件大小 */
+        uint32_t    filesize;   /**< 文件大小 */
     };
 }
